+++ src/main/test/edu/fiuba/algo3/otros_casos_de_uso/ObstaculoAsisteAUnBacanalTest.java
@@ -1,26 +1,27 @@
-package edu.fiuba.algo3.entrega_1.otros;
+package edu.fiuba.algo3.otros_casos_de_uso;
 
 import edu.fiuba.algo3.modelo.casilleros.Casillero;
+import edu.fiuba.algo3.modelo.consecuencias.AsisteAUnBacanal;
 import edu.fiuba.algo3.modelo.consecuencias.Consecuencia;
-import edu.fiuba.algo3.modelo.consecuencias.Lesion;
 import edu.fiuba.algo3.modelo.equipamientos.Equipamiento;
 import edu.fiuba.algo3.modelo.equipamientos.SinEquipamiento;
 import edu.fiuba.algo3.modelo.gladiador.Energia;
 import edu.fiuba.algo3.modelo.gladiador.Gladiador;
 import edu.fiuba.algo3.modelo.gladiador.senority.Novato;
 import edu.fiuba.algo3.modelo.gladiador.senority.Senority;
-import edu.fiuba.algo3.modelo.MovimientoPausadoExeption;
+import edu.fiuba.algo3.modelo.MovimientoException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class ObstaculoLesionTest {
+public class ObstaculoAsisteAUnBacanalTest {
 
     @Test
-    public void verificarQueSiRecibeUnObstaculoLesionYSeEncuentraEnELTurnoSiguienteNoSePuedeMover() throws Exception {
+    public void verificarQueSiRecibeUnObstaculoAsisteAUnBacanalSacaCuatroPuntosDeEnergiaPorCadaTragoTomado() throws Exception {
         //Arrange
-        Consecuencia lesion = new Lesion();
+        int cantiadDeTragos = 6;
+        Consecuencia asisteAUnBacanal = new AsisteAUnBacanal(cantiadDeTragos);
         Energia energia = new Energia(20);
         Equipamiento equipamiento = new SinEquipamiento();
         Casillero casillero = new Casillero(0);
@@ -28,12 +29,12 @@ public class ObstaculoLesionTest {
         Gladiador gladiador = new Gladiador(energia, equipamiento, casillero, senority);
 
         //Act
-        gladiador.recibirConsecuencia(lesion);
+        gladiador.recibirConsecuencia(asisteAUnBacanal);
 
         //Assert
-        Throwable exception= Assertions.assertThrows(MovimientoPausadoExeption.class, () -> {
+        Throwable exception= Assertions.assertThrows(MovimientoException.class, () -> {
             gladiador.avanzar(1);
         });
-        assertEquals("El gladiador esta pausado para mover en este turno", exception.getMessage());
+        assertEquals("El gladiador no se puede mover sin energia", exception.getMessage());
     }
 }
